fixed part 4 depencies

+++ Part 4/controllers/blogs.js
@@ -1,44 +1,62 @@
 //route handlers
 const router = require('express').Router()
+const jwt = require('jsonwebtoken')
+
 const Blog = require('../models/blog')
+const User = require('../models/user')
 
 router.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
-   response.json(blogs)
-  })
+  const notes = await Blog
+    .find({})
+    .find({}).populate('user', { username: 1, name: 1 })
+
+  response.json(notes)
+})
 
 router.post('/', async (request, response) => {
-  const body = request.body
-  const blog = new Blog({
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    likes: body.likes,
-  })
+  if (!request.user) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const user = request.user
+  const blog = new Blog({ ...request.body, user: user.id })
+
   const savedBlog = await blog.save()
-    response.json(savedBlog)
-    })
+
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
+  response.status(201).json(savedBlog)
+})
 
 router.delete('/:id', async (request, response) => {
+  const blogToDelete = await Blog.findById(request.params.id)
+  if (!blogToDelete ) {
+    return response.status(204).end()
+  }
+
+  if ( blogToDelete.user && blogToDelete.user.toString() !== request.user.id ) {
+    return response.status(401).json({
+      error: 'only the creator can delete a blog'
+    })
+  }
+
   await Blog.findByIdAndRemove(request.params.id)
-      response.status(204).end()
+
+  response.status(204).end()
 })
 
 router.put('/:id', async (request, response) => {
-  const body = request.body
-  const blog = {
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    likes: body.likes
-  }
+  const blog = request.body
+
   const updatedBlog = await Blog
     .findByIdAndUpdate(
       request.params.id, 
       blog, 
-      { new: true })
+      { new: true, runValidators: true, context: 'query' }
+    )
       
-    response.json(updatedBlog)
+  response.json(updatedBlog)
 })
 
 module.exports = router
